requirements.txt
discord.py
langchain
faiss-cpu
psycopg2-binary
flask
python-dotenv
openai
